# specs/002-web-dashboard/openapi.yaml
openapi: 3.0.3
info:
  title: Lukas the Bear Admin Dashboard API
  description: |
    REST API for the Lukas the Bear Admin Dashboard. Provides monitoring and control
    capabilities for the Slack chatbot including activity logs, generated images,
    scheduled events, and manual control triggers.

    ## Authentication
    All endpoints (except login) require a valid session cookie obtained through
    the POST /api/auth/login endpoint. Sessions remain valid until the browser is closed.

    ## Rate Limiting
    Manual control endpoints are rate-limited to prevent abuse (max 10 requests per hour per session).
  version: 1.0.0
  contact:
    name: Admin Dashboard Support
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://dashboard.lukas-bear.example.com
    description: Production server
...
paths:
...
  /api/activity:
    get:
      tags:
        - Activity Logs
      summary: Get paginated activity log
      description: |
        Returns a paginated list of messages sent by Lukas the Bear, with optional filtering
        by date range, recipient, and channel type. Results are ordered by timestamp descending
        (most recent first).
      operationId: getActivityLog
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          example: 50
        - name: start_date
          in: query
          description: Filter messages from this date onwards (ISO 8601 format)
          schema:
            type: string
            format: date-time
          example: "2025-10-01T00:00:00Z"
        - name: end_date
          in: query
          description: Filter messages until this date (ISO 8601 format)
          schema:
            type: string
            format: date-time
          example: "2025-10-28T23:59:59Z"
        - name: recipient
          in: query
          description: Filter by recipient Slack user ID
          schema:
            type: string
          example: "U01234ABCDE"
        - name: channel_type
          in: query
          description: Filter by channel type
          schema:
            type: string
            enum:
              - dm
              - channel
              - thread
          example: "dm"
      responses:
        '200':
          description: Activity log retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityLogEntry'
                  total:
                    type: integer
                    description: Total number of matching items
                    example: 247
                  page:
                    type: integer
                    description: Current page number
                    example: 1
                  pages:
                    type: integer
                    description: Total number of pages
                    example: 5
              examples:
                success:
                  value:
                    items:
                      - id: "msg-123"
                        conversation_id: "conv-456"
                        timestamp: "2025-10-28T15:30:00Z"
                        recipient_user_id: "U01234ABCDE"
                        recipient_display_name: "John Doe"
                        channel_type: "dm"
                        channel_id: "D987654321"
                        content_preview: "Hey! I hope you're having a great day! Want to chat about..."
                        token_count: 42
                    total: 247
                    page: 1
                    pages: 5
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/activity/{id}:
    get:
      tags:
        - Activity Logs
      summary: Get full message details
      description: |
        Returns complete details for a specific message including full content,
        conversation context, and any tool invocations that occurred.
      operationId: getActivityDetail
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Message ID
          schema:
            type: string
          example: "msg-123"
      responses:
        '200':
          description: Message details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetail'
              examples:
                success:
                  value:
                    id: "msg-123"
                    conversation_id: "conv-456"
                    timestamp: "2025-10-28T15:30:00Z"
                    recipient_user_id: "U01234ABCDE"
                    recipient_display_name: "John Doe"
                    channel_type: "dm"
                    channel_id: "D987654321"
                    content: "Hey! I hope you're having a great day! Want to chat about your weekend plans?"
                    token_count: 42
                    metadata:
                      tool_invocations: []
                      model_used: "claude-3-5-sonnet-20241022"
                    conversation_context:
                      message_count: 5
                      last_message_at: "2025-10-28T15:30:00Z"
                      context_summary: "Friendly conversation about weekend plans"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

  # Generated Images Endpoints
...
  /api/events/upcoming:
    get:
      tags:
        - Scheduled Events
      summary: Get upcoming scheduled events
      description: |
        Returns scheduled tasks that are pending execution, ordered by scheduled time
        (soonest first).
      operationId: getUpcomingEvents
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of events to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          example: 50
      responses:
        '200':
          description: Upcoming events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduledEvent'
              examples:
                success:
                  value:
                    items:
                      - id: "task-101"
                        job_id: "job_random_dm_20251028_153000"
                        task_type: "random_dm"
                        target_type: "user"
                        target_id: "U01234ABCDE"
                        scheduled_at: "2025-10-28T15:30:00Z"
                        executed_at: null
                        status: "pending"
                        retry_count: 0
                        error_message: null
                        metadata:
                          trigger_source: "automatic"
                        created_at: "2025-10-28T08:00:00Z"
                      - id: "task-102"
                        job_id: "job_image_post_20251028_180000"
                        task_type: "image_post"
                        target_type: "channel"
                        target_id: "C01234567"
                        scheduled_at: "2025-10-28T18:00:00Z"
                        executed_at: null
                        status: "pending"
                        retry_count: 0
                        error_message: null
                        metadata:
                          theme: "seasonal_autumn"
                          trigger_source: "automatic"
                        created_at: "2025-10-28T08:00:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/events/completed:
    get:
      tags:
        - Scheduled Events
      summary: Get completed/historical events
      description: |
        Returns paginated list of scheduled tasks that have been executed, failed, or cancelled,
        ordered by execution timestamp descending (most recent first).
      operationId: getCompletedEvents
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          example: 50
      responses:
        '200':
          description: Completed events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduledEvent'
                  total:
                    type: integer
                    description: Total number of matching items
                    example: 156
                  page:
                    type: integer
                    description: Current page number
                    example: 1
                  pages:
                    type: integer
                    description: Total number of pages
                    example: 4
              examples:
                success:
                  value:
                    items:
                      - id: "task-100"
                        job_id: "job_random_dm_20251027_140000"
                        task_type: "random_dm"
                        target_type: "user"
                        target_id: "U98765ZYXWV"
                        scheduled_at: "2025-10-27T14:00:00Z"
                        executed_at: "2025-10-27T14:00:03Z"
                        status: "completed"
                        retry_count: 0
                        error_message: null
                        metadata:
                          trigger_source: "automatic"
                          message_sent_ts: "1698417603.123456"
                        created_at: "2025-10-27T08:00:00Z"
                      - id: "task-99"
                        job_id: "job_image_post_20251026_120000"
                        task_type: "image_post"
                        target_type: "channel"
                        target_id: "C01234567"
                        scheduled_at: "2025-10-26T12:00:00Z"
                        executed_at: "2025-10-26T12:00:15Z"
                        status: "failed"
                        retry_count: 3
                        error_message: "OpenAI API error: Rate limit exceeded"
                        metadata:
                          theme: "celebration"
                          trigger_source: "automatic"
                        created_at: "2025-10-26T08:00:00Z"
                    total: 156
                    page: 1
                    pages: 4
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

  # Manual Control Endpoints
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie obtained from POST /api/auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Short error identifier
          example: "Unauthorized"
        message:
          type: string
          description: Human-readable error message
          example: "Authentication required"
      required:
        - error
        - message
    ActivityLogEntry:
      type: object
      description: Summary of a bot message for activity log listing
      properties:
        id:
          type: string
          description: Unique message ID
          example: "msg-123"
        conversation_id:
          type: string
          description: ID of the conversation this message belongs to
          example: "conv-456"
        timestamp:
          type: string
          format: date-time
          description: When the message was sent (ISO 8601)
          example: "2025-10-28T15:30:00Z"
        recipient_user_id:
          type: string
          description: Slack user ID of the recipient
          example: "U01234ABCDE"
        recipient_display_name:
          type: string
          description: Display name of the recipient
          example: "John Doe"
        channel_type:
          type: string
          enum:
            - dm
            - channel
            - thread
          description: Type of channel where message was sent
          example: "dm"
        channel_id:
          type: string
          nullable: true
          description: Slack channel ID (null for DMs)
          example: "D987654321"
        content_preview:
          type: string
          description: First 100 characters of message content
          example: "Hey! I hope you're having a great day! Want to chat about..."
        token_count:
          type: integer
          description: Number of tokens used for this message
          example: 42
      required:
        - id
        - conversation_id
        - timestamp
        - recipient_user_id
        - channel_type
        - content_preview
        - token_count
    ActivityDetail:
      type: object
      description: Full details of a bot message
      properties:
        id:
          type: string
          description: Unique message ID
          example: "msg-123"
        conversation_id:
          type: string
          description: ID of the conversation this message belongs to
          example: "conv-456"
        timestamp:
          type: string
          format: date-time
          description: When the message was sent (ISO 8601)
          example: "2025-10-28T15:30:00Z"
        recipient_user_id:
          type: string
          description: Slack user ID of the recipient
          example: "U01234ABCDE"
        recipient_display_name:
          type: string
          description: Display name of the recipient
          example: "John Doe"
        channel_type:
          type: string
          enum:
            - dm
            - channel
            - thread
          description: Type of channel where message was sent
          example: "dm"
        channel_id:
          type: string
          nullable: true
          description: Slack channel ID (null for DMs)
          example: "D987654321"
        content:
          type: string
          description: Full message content
          example: "Hey! I hope you're having a great day! Want to chat about your weekend plans?"
        token_count:
          type: integer
          description: Number of tokens used for this message
          example: 42
        metadata:
          type: object
          nullable: true
          description: Additional message metadata (tool invocations, model used, etc.)
          example:
            tool_invocations: []
            model_used: "claude-3-5-sonnet-20241022"
        conversation_context:
          type: object
          description: Context about the conversation this message belongs to
          properties:
            message_count:
              type: integer
              description: Total messages in this conversation
              example: 5
            last_message_at:
              type: string
              format: date-time
              description: Timestamp of most recent message in conversation
              example: "2025-10-28T15:30:00Z"
            context_summary:
              type: string
              nullable: true
              description: Brief summary of conversation context
              example: "Friendly conversation about weekend plans"
      required:
        - id
        - conversation_id
        - timestamp
        - recipient_user_id
        - channel_type
        - content
        - token_count
...
    ScheduledEvent:
      type: object
      description: A scheduled or completed task
      properties:
        id:
          type: string
          description: Unique task ID
          example: "task-101"
        job_id:
          type: string
          description: APScheduler job ID
          example: "job_random_dm_20251028_153000"
        task_type:
          type: string
          enum:
            - random_dm
            - image_post
            - cleanup
            - reminder
            - manual_image
            - manual_dm
          description: Type of scheduled task
          example: "random_dm"
        target_type:
          type: string
          enum:
            - user
            - channel
            - system
          description: Type of target for the task
          example: "user"
        target_id:
          type: string
          nullable: true
          description: Slack user ID or channel ID (null for system tasks)
          example: "U01234ABCDE"
        scheduled_at:
          type: string
          format: date-time
          description: When the task is/was scheduled to execute (ISO 8601)
          example: "2025-10-28T15:30:00Z"
        executed_at:
          type: string
          format: date-time
          nullable: true
          description: When the task was executed (ISO 8601, null if not executed)
          example: "2025-10-28T15:30:03Z"
        status:
          type: string
          enum:
            - pending
            - executing
            - completed
            - failed
            - cancelled
          description: Current status of the task
          example: "completed"
        retry_count:
          type: integer
          description: Number of retry attempts made
          example: 0
        error_message:
          type: string
          nullable: true
          description: Error message if task failed
          example: null
        metadata:
          type: object
          nullable: true
          description: Task-specific metadata (theme, trigger source, result timestamps)
          example:
            trigger_source: "manual"
            theme: "celebration"
            message_sent_ts: "1698494490.123456"
        created_at:
          type: string
          format: date-time
          description: When the task record was created (ISO 8601)
          example: "2025-10-28T08:00:00Z"
      required:
        - id
        - job_id
        - task_type
        - target_type
        - scheduled_at
        - status
        - retry_count
        - created_at
...
  responses:
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unauthorized:
              value:
                error: "Unauthorized"
                message: "Authentication required. Please log in."
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notFound:
              value:
                error: "Not found"
                message: "The requested resource does not exist"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            serverError:
              value:
                error: "Internal server error"
                message: "An unexpected error occurred. Please try again later."